Change quantity to be a float

+++ db/migrate/20200308041155_create_snack_events.rb
@@ -4,7 +4,7 @@ class CreateSnackEvents < ActiveRecord::Migration[5.2]
   def change
     create_table :snack_events do |t|
       t.datetime :event_time, null: false
-      t.integer :quantity, null: false
+      t.float :quantity, null: false
       t.references :victual_item, foreign_key: true, null: false
     end
   end
